Access user data automatically if an OAuth2 login flow does not provide any.
The Authomatic API generally returns a populated User object from a login
authentication flow, except for OAuth2 providers (e.g. Facebook) for which
an addition user data lookup request is required.

With this change, a User data update is performed automatically if necessary
to populate the User object with useful information.